fix: suppress clippy unnecessary_cast for cross-platform statvfs fields
f_bavail and f_frsize types differ between Linux (u64) and macOS (u32).
The `as u64` cast is needed on macOS but redundant on Linux. Suppress
the clippy lint with #[allow] for this cross-platform line.

+++ native/ferricstore_bitcask/src/store.rs
@@ -783,7 +783,8 @@ fn available_disk_space_for_path(path: &Path) -> Result<u64> {
                     std::io::Error::last_os_error()
                 )));
             }
-            Ok(stat.f_bavail as u64 * stat.f_frsize)
+            #[allow(clippy::unnecessary_cast)]
+            Ok(stat.f_bavail as u64 * stat.f_frsize as u64)
         }
     }
     #[cfg(not(unix))]
